fix: the fault by expanding express.request
