+++ backend/nodemon.json
@@ -1,0 +1,10 @@
+{
+  "watch": ["src"],
+  "ext": "ts,json",
+  "ignore": ["src/**/*.spec.ts", "src/**/*.test.ts"],
+  "execMap": {
+    "ts": "ts-node -r tsconfig-paths/register"
+  },
+  "exec": "ts-node -r tsconfig-paths/register src/server.ts",
+  "verbose": true
+}
