+++ backend/src/shared/middleware/validate-schema.ts
@@ -1,6 +1,6 @@
 import z from 'zod';
 import { NextFunction, RequestHandler, Response } from 'express';
-import { TypedRequest } from '../../types/app-express/types';
+import { TypedRequest } from '../../types/express-interfaces/types';
 import { ResponseHelper } from '../utils/response.helpers';
 
 export const validateSchema = <T extends z.ZodTypeAny>(schema: T): RequestHandler => {
