--- backend/src/types/app-express/index.d.ts
@@ -1,16 +1,0 @@
-import 'express';
-
-declare global {
-  namespace Express {
-    interface User {
-      id: string;
-      username: string;
-      role: string;
-      roles: string[];
-    }
-
-    interface Request {
-      id: string;
-    }
-  }
-}
