+++ backend/src/types/express.d.ts
@@ -1,0 +1,13 @@
+declare namespace Express {
+  interface User {
+    id: string;
+    username: string;
+    role: string;
+    roles: string[];
+  }
+
+  interface Request {
+    id: string;
+    user?: User;
+  }
+}
